code refactor and optimization
